Working script finally
Updated the script in the paper and got it working. Currently imports arbitrary data, runs the parameter fit, and then time steps Cases and Deaths with the optimized parameters to produce a very good fit for the data. Also plots cases against numerical best fit cases, and same for deaths.